feat: move form to its own component
