+++ src/components/TaskList.js
@@ -1,33 +1,14 @@
 import Task from './Task';
-import { useInput } from '../hooks/useInput';
+import TaskForm from './TaskForm';
 
 const TaskList = ({ tasks, updateTask, addTask }) => {
-  const [{ task }, update, reset] = useInput({ task: '' });
-
-  const handleInputChange = (event) => {
-    update(event);
-  };
-
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    addTask(task);
-    reset();
-  };
-
   return (
     <ul>
       {tasks.map((task) => (
         <Task key={task.id} task={task} updateTask={updateTask} />
       ))}
       <li>
-        <form onSubmit={handleSubmit}>
-          <input
-            type="text"
-            name="task"
-            value={task}
-            onChange={handleInputChange}
-          />
-        </form>
+        <TaskForm addTask={addTask} />
       </li>
     </ul>
   );
